DualIndicatorPushbutton, IndicatorPushButton and Indicator corrected to restore xml font size after the base.ReadXml.  Pushbutton has been altered to correctly deal with scaling text, and the newish Rocker with Text gets scaling.
