+++ Helios/Controls/DualIndicatorPushButtonAppearanceEditor.xaml
@@ -102,7 +102,7 @@
         </Label>
         <Slider Grid.Row="3" Grid.Column="1" Minimum="0.1" Maximum="1" Value="{Binding Control.GlyphScale}" Margin="4,4,4,4" TickPlacement="TopLeft" TickFrequency="0.1" Style="{StaticResource SliderGlyphStyle}"/>
         <TextBlock Grid.Column="0" Grid.ColumnSpan="2" Grid.Row="4" Margin="4,10,2,2" Style="{StaticResource Heading2}" Text="Text Properties"/>
-        <Label Grid.Row="5" FontSize="10" HorizontalAlignment="Center" VerticalAlignment="Top" Content="Auto Scaling">
+        <Label Grid.Row="5" FontSize="10" HorizontalAlignment="Center" VerticalAlignment="Top" Content="Text Scaling">
             <Label.ToolTip>How font size will be scaled if the control is scaled.</Label.ToolTip>
         </Label>
         <ComboBox Grid.Column="1" Grid.Row="5" Margin="2,2,0,2" HorizontalAlignment="Left" ItemsSource="{Binding Source={StaticResource TextScalingModes}, Path=DisplayNames}" SelectedItem="{Binding Path=Control.ScalingMode, Converter={StaticResource TextScalingModes}}"/>
