+++ Helios/Controls/PushButton.cs
@@ -16,6 +16,7 @@
 namespace GadrocsWorkshop.Helios.Controls
 {
     using GadrocsWorkshop.Helios.ComponentModel;
+    using NLog;
     using System;
     using System.ComponentModel;
     using System.Globalization;
@@ -53,6 +54,8 @@ namespace GadrocsWorkshop.Helios.Controls
         private HeliosValue _value;
         private HeliosValue _pushedValue;
 
+        private static readonly NLog.Logger Logger = NLog.LogManager.GetCurrentClassLogger();
+
         public PushButton()
             : base("Push Button", new Size(50, 50))
         {
@@ -72,6 +75,8 @@ namespace GadrocsWorkshop.Helios.Controls
             _labelFormat.FontSize = 20;
             _labelFormat.HorizontalAlignment = TextHorizontalAlignment.Center;
             _labelFormat.VerticalAlignment = TextVerticalAlignment.Center;
+            _referenceHeight = Height;
+            _scalingMode = TextScalingMode.Height;
 
             _pushedTrigger = new HeliosTrigger(this, "", "", "pushed", "Fired when this button is pushed.", "Always returns true.", BindingValueUnits.Boolean);
             _releasedTrigger = new HeliosTrigger(this, "", "", "released", "Fired when this button is released.", "Always returns false.", BindingValueUnits.Boolean);
@@ -319,6 +324,7 @@ namespace GadrocsWorkshop.Helios.Controls
             get { return _labelFormat; }
         }
 
+
         public Point TextPushOffset
         {
             get
@@ -336,7 +342,31 @@ namespace GadrocsWorkshop.Helios.Controls
                 }
             }
         }
+        /// <summary>
+        /// backing field for property ScalingMode, contains
+        /// the selected automatic font size scaling mode
+        /// </summary>
+        private TextScalingMode _scalingMode;
 
+        /// <summary>
+        /// the height this display had when the font size was configured
+        /// </summary>
+        private double _referenceHeight;
+
+        /// <summary>
+        /// the selected automatic font size scaling mode
+        /// </summary>
+        public TextScalingMode ScalingMode
+        {
+            get => _scalingMode;
+            set
+            {
+                if (_scalingMode == value) return;
+                TextScalingMode oldValue = _scalingMode;
+                _scalingMode = value;
+                OnPropertyChanged("ScalingMode", oldValue, value, true);
+            }
+        }
         #endregion
 
         void Format_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
@@ -499,7 +529,20 @@ namespace GadrocsWorkshop.Helios.Controls
                 {
                     TextPushOffset = (Point)pointConverter.ConvertFromInvariantString(reader.ReadElementString("TextPushOffset"));
                 }
-                reader.ReadEndElement();
+                if (reader.Name.Equals("ScalingMode") && Enum.TryParse(reader.ReadElementString("ScalingMode"), out TextScalingMode configured))
+                {
+                    ScalingMode = configured;
+                }
+                else
+                {
+                    ScalingMode = TextScalingMode.Legacy;
+                }
+                
+            reader.ReadEndElement();
+            // now the auto scaling has messed up our font size, so we restore it
+            TextFormat.FontSize = TextFormat.ConfiguredFontSize;
+            _referenceHeight = Height;
+
             }
         }
 
@@ -539,6 +582,10 @@ namespace GadrocsWorkshop.Helios.Controls
                 {
                     writer.WriteElementString("TextPushOffset", pointConverter.ConvertToString(null, CultureInfo.InvariantCulture, TextPushOffset));
                 }
+                if (ScalingMode != TextScalingMode.Legacy)
+                {
+                    writer.WriteElementString("ScalingMode", ScalingMode.ToString());
+                }
                 writer.WriteEndElement();
             }
         }
@@ -547,13 +594,59 @@ namespace GadrocsWorkshop.Helios.Controls
 
         public override void ScaleChildren(double scaleX, double scaleY)
         {
+            //if (GlobalOptions.HasScaleAllText)
+            //{
+            //    _labelFormat.FontSize *= Math.Max(scaleX, scaleY);
+            //}
+            base.ScaleChildren(scaleX, scaleY);
+        }
+        // WARNING: this virtual method is called from the base constructor (indirectly)
+        protected override void PostUpdateRectangle(Rect previous, Rect current)
+        {
             if (GlobalOptions.HasScaleAllText)
             {
-                _labelFormat.FontSize *= Math.Max(scaleX, scaleY);
+                switch (ScalingMode)
+                {
+                    case TextScalingMode.Height:
+                        if (_referenceHeight < 0.001)
+                        {
+                            _labelFormat.FontSize = _labelFormat.ConfiguredFontSize;
+                            break;
+                        }
+                        // avoid accumulating error from repeated resizing by calculating from a reference point
+                        Logger.Debug("scaling font based on new height {Height} versus reference {ReferenceSize} at height {ReferenceHeight}",
+                            current.Height, _labelFormat.ConfiguredFontSize, _referenceHeight);
+                        _labelFormat.FontSize = Clamp(_labelFormat.ConfiguredFontSize * current.Height / _referenceHeight, 1, 2000);
+                        break;
+                    case TextScalingMode.None:
+                        return;
+                    case TextScalingMode.Legacy:
+                        if (previous.Height != 0)
+                        {
+                            double scale = current.Height / previous.Height;
+                            _labelFormat.FontSize = Clamp(scale * _labelFormat.FontSize, 1, 100);
+                        }
+                        break;
+                    default:
+                        throw new ArgumentOutOfRangeException();
+                }
+                Logger.Debug("Font Size " + _labelFormat.FontSize);
             }
-            base.ScaleChildren(scaleX, scaleY);
         }
 
         #endregion
+        private double Clamp(double value, double min, double max)
+        {
+            if (value < min)
+            {
+                return min;
+            }
+            if (value > max)
+            {
+                return max;
+            }
+            return value;
+        }
+
     }
 }
